+++ app/Http/Controllers/ReturnPurchaseController.php
@@ -101,13 +101,13 @@ class ReturnPurchaseController extends Controller
             ])
             ->whereNotNull('product_warehouse.variant_id')
             ->get();
-        
+
         $product_code = [];
         $product_name = [];
         $product_qty = [];
         $is_batch = [];
         $product_data = [];
-        foreach ($lims_product_warehouse_data as $product_warehouse) 
+        foreach ($lims_product_warehouse_data as $product_warehouse)
         {
             $product_qty[] = $product_warehouse->qty;
             $product_code[] =  $product_warehouse->code;
@@ -116,7 +116,7 @@ class ReturnPurchaseController extends Controller
             $is_batch[] = null;
         }
         //product with batches
-        foreach ($lims_product_with_batch_warehouse_data as $product_warehouse) 
+        foreach ($lims_product_with_batch_warehouse_data as $product_warehouse)
         {
             $product_qty[] = $product_warehouse->qty;
             $lims_product_data = Product::select('code', 'name', 'type', 'is_batch')->find($product_warehouse->product_id);
@@ -127,7 +127,7 @@ class ReturnPurchaseController extends Controller
             $is_batch[] = $lims_product_data->is_batch;
         }
 
-        foreach ($lims_product_with_variant_warehouse_data as $product_warehouse) 
+        foreach ($lims_product_with_variant_warehouse_data as $product_warehouse)
         {
             $lims_product_variant_data = ProductVariant::select('item_code')->FindExactProduct($product_warehouse->id, $product_warehouse->variant_id)->first();
             $product_qty[] = $product_warehouse->qty;
@@ -159,7 +159,7 @@ class ReturnPurchaseController extends Controller
         $product[] = $lims_product_data->name;
         $product[] = $lims_product_data->code;
         $product[] = $lims_product_data->cost;
-        
+
         if ($lims_product_data->tax_id) {
             $lims_tax_data = Tax::find($lims_product_data->tax_id);
             $product[] = $lims_tax_data->rate;
@@ -188,7 +188,7 @@ class ReturnPurchaseController extends Controller
                 $unit_operation_value[] = $unit->operation_value;
             }
         }
-        
+
         $product[] = implode(",", $unit_name) . ',';
         $product[] = implode(",", $unit_operator) . ',';
         $product[] = implode(",", $unit_operation_value) . ',';
@@ -224,7 +224,7 @@ class ReturnPurchaseController extends Controller
         $lims_return_data = ReturnPurchase::create($data);
         if($data['supplier_id']){
             $lims_supplier_data = Supplier::find($data['supplier_id']);
-        
+
             //collecting male data
             $mail_data['email'] = $lims_supplier_data->email;
             $mail_data['reference_no'] = $lims_return_data->reference_no;
@@ -400,7 +400,7 @@ class ReturnPurchaseController extends Controller
         }
         else
             $message = "This return doesn't belong to any supplier";
-        
+
         return redirect()->back()->with('message', $message);
     }
 
@@ -436,7 +436,7 @@ class ReturnPurchaseController extends Controller
             );
             if ($v->fails())
                 return redirect()->back()->withErrors($v->errors());
-            
+
             $documentName = $document->getClientOriginalName();
             $document->move('public/return/documents', $documentName);
             $data['document'] = $documentName;
@@ -598,7 +598,7 @@ class ReturnPurchaseController extends Controller
             $mail_data['order_tax'] = $lims_return_data->order_tax;
             $mail_data['order_tax_rate'] = $lims_return_data->order_tax_rate;
             $mail_data['grand_total'] = $lims_return_data->grand_total;
-            
+
             try{
                 Mail::send( 'mail.return_details', $mail_data, function( $message ) use ($mail_data)
                 {
@@ -656,7 +656,10 @@ class ReturnPurchaseController extends Controller
                     $product_return_data->delete();
                 }
             }
-            $lims_return_data->delete();
+
+            if($lims_return_data){
+                $lims_return_data->delete();
+            }
         }
         return 'Return deleted successfully!';
     }
@@ -703,7 +706,11 @@ class ReturnPurchaseController extends Controller
                 $product_return_data->delete();
             }
         }
-        $lims_return_data->delete();
+
+        if($lims_return_data){
+            $lims_return_data->delete();
+        }
+
         return redirect('return-purchase')->with('not_permitted', 'Data deleted successfully');;
     }
 }
